# Implement LinearTime algorithm for finding approximate maximum independent set