This implements the "LinearTime" algorithm proposed in [1]. It is capable of finding a large independent set in linear time.
However, there is no guarantee for the ratio between the size of the returned set and the actual maximum independent set of the graph.

My implementation is partly based on the "reference" implementation [2].

In the current draft, the implementation completely replaces the current functions for `nx.approximation.maximum_independent_set` and `nx.approximation.max_clique`. We may want to offer both options instead.

### Motivation

The current implementations of approximate maximum independent sets and maximum cliques are implemented recursively based on a 1992 paper [3] and either take ages, run out of memory or hit maximum recursion depth (MRD) for moderately sized graphs. This new implementation runs in linear time and is thus extremely fast even on graphs with multiple millions of edges (see benchmark results).

### Benefits

This implementation:
- runs in linear time O(n+m), assuming dict and set accesses to be constant
- does not copy the graph but instead temporarily modifies it
- offers good empirical results [1]

### Benchmarks
Graphs:
| Name | n | m| Source |
|--------|--------:|--------:|--------:|
| GrQc | 5,242 | 14,484 | [SNAP](https://snap.stanford.edu/data/ca-GrQc.html) |
| CondMat | 23,133 | 93,439 | [SNAP](https://snap.stanford.edu/data/ca-CondMat.html) |
| AstroPh | 18,772 | 198,050 | [SNAP](https://snap.stanford.edu/data/ca-AstroPh.html) |
| Genes | 1,678 | 31,944 | personal work | 

The edge count is without self-loops for each graph.

Results:
| Algorithm | GrQc | CondMat | AstroPh | Genes |
|--------|--------|--------|--------|--------|
| **current** | MRD (10s) | MRD (85s) | MRD (121s) | 86 (70s) |
| **linear** | 2,459 (0.02s) | 9,611 (0.18s) | 6,760 (0.21s) | 73 (0.03s) || 

These benchmarks were run on a Framework 13 (AMD Ryzen™ 5 7640U, 16 GB RAM) with around 5GB of RAM available.
For the cases where the benchmark failed to produce a result, the time measurements include the time needed to load the graph from disk.

On [BerkStan](https://snap.stanford.edu/data/web-BerkStan.html) (n=685,230, m=6,649,470), linear produces 407,703 (779 from opt) in ~10s. I would appreciate if someone could test the quality of linear's results compared to the current implementation, e.g. with random small graphs.

### Caveats
The algorithm:
- offers no proven guarantee for the quality of the results
- can perform worse than the current algorithm quality-wise on small graphs

Because of the poor performance on large graphs, users currently probably use these functions for relatively small graphs and could thus be surprised if the quality of the results suddenly decreases. On the other hand, it's an approximation of an usually unknown value so giving worse results but much faster may be a worth trade-off.

### Open questions
- Is my implementation fine this way? (naming, documentation, etc.)
- How should randomness be handled? Is a seed preferable?
- How should we integrate this? (replace current functions vs fast-flag vs alternate functions)

### Future direction
The paper [1] suggests a second algorithm "NearLinear" that promises better results than "NearLinear" for the trade-off of running in O(m x d) where d is the maximum degree of the graph. I have implemented a draft of this algorithm, however it doesn't appear to be correct (suboptimal results despite only using exact reductions) and I am slightly disappointed with the runtime as well, though I have not implemented the lp reduction. If there is an interest in the code, let me know and we might be able to fix the bugs together.

As the max_clique approximation performs the independent_set approximation on the complement of the graph, it is faster for dense graphs than for sparse graphs. We might want to add an additional heuristic algorithm for an approximate maximum clique on sparse graphs and decide which one to use based on the edge count. 

### References
---
[1] Chang, Lijun, Wei Li, and Wenjie Zhang.
"Computing a Near-Maximum Independent Set in Linear Time by Reducing-Peeling."
Proceedings of the 2017 ACM International Conference on Management of Data (2017): 1181–96.
[Researchgate link](https://www.researchgate.net/profile/Wei-Li-291/publication/316849563_Computing_A_Near-Maximum_Independent_Set_in_Linear_Time_by_Reducing-Peeling/).
[2] https://github.com/LijunChang/Near-Maximum-Independent-Set/
[3] Boppana, R., & Halldórsson, M. M. (1992).
Approximating maximum independent sets by excluding subgraphs.
BIT Numerical Mathematics, 32(2), 180–196. Springer.