@maxtidev 2025-07-24
> There is precedent for a "method" kwarg that allows users to choose between implementations

I went ahead and implemented this for `approximation.maximum_independent_set` and `approximation.max_clique` as I really like that approach. I also updated the tests and documentation to reflect both implemented methods.

This would also make it easy to add the NearLinear algorithm if I or someone else ever feels like implementing that too.