@maxtidev 2025-08-04
Alright, thank you for the advice!  
The last open question for me is non-determinism. Testing multiple times on the "genes" graph, I get different results (between size 71 and 75) and I am unsure which part of the algorithm is responsible for this randomness.
It is not `maximal_independent_set()`, so much I was able to figure out. Does anyone have an idea what might be the culprit and how to avoid it?  
Even if the algorithm is just an approximation, I think it would be nice if it was deterministic given that it's not a randomized algorithmic approach.