@boothby 2025-08-08
> The last open question for me is non-determinism.

The ordering of elements in sets is non-deterministic; if you really need
determinism that's easiest to enforce by using a dict instead of a set for
the members of `self.nodes_by_degree`

On Thu, Aug 7, 2025 at 11:18 AM Max T. ***@***.***> wrote:

> ***@***.**** commented on this pull request.
> ------------------------------
>
> In networkx/algorithms/approximation/clique.py
> <https://github.com/networkx/networkx/pull/8159#discussion_r2261101718>:
>
> >      """
>
>      def __init__(self, G):
>          """Initialize the class. Self-loop edges are temporarily removed."""
>          self.G = G
> -        if G.order() == 0:
> +
> +        # note: no edge or node data is copied
> +        self.H = nx.Graph(G.edges)
>
> We need self.G in lines 205-217
> <https://github.com/networkx/networkx/pull/8159/files/6da5eed5782329eb41471fa17775362bcbb641c2#diff-e97f086902ba4c29badd770fafdf137d89ee6e71e54951300858bd4ab413d498R205-R217>
> to extend the independent set using the input graph's adjacency
> information. Since we delete nodes and their edges from self.H, that copy
> can't be used at that stage of the algorithm.
>
> Since we still have to temporarily remove self-loops from self.G there,
> it would also be possible to operate on G in the __init__ function and
> copy G *afterwards*. Is there really a noticable benefit to reducing
> self.H lookups? Because this alternative implementation would add a class
> variable just for the self-loops to be added back later and I'm unsure
> whether that's cleaner.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/networkx/networkx/pull/8159#discussion_r2261101718>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAELCNX6YT4IIRIHMOUYQK33MOKAFAVCNFSM6AAAAACCAQQIUSVHI2DSMVQWIX3LMV43YUDVNRWFEZLROVSXG5CSMV3GSZLXHMZTAOJYGI4TQNBXGM>
> .
> You are receiving this because you are subscribed to this thread.Message
> ID: ***@***.***>
>