@rossbar 2025-08-13
> Thanks! This does work, but it definitely hurts readability (see https://github.com/networkx/networkx/issues/3912) and complicates some parts that would otherwise be simple set subtractions.

I think what @boothby is that Python `dict` s are both ordered and support set operations:

```python
>>> d = dict(enumerate(range(10)))
>>> d.keys() - {1, 3, 5}
{0, 2, 4, 6, 7, 8, 9}
```

As of Python 3.7, dicts are also guaranteed to be ordered, which essentially provides an ordered set. This wasn't universally the case when #3912 was opened but the issue is no longer relevant (IMO) given these features which are now part of Python itself.