@maxtidev 2025-08-14
I wasn't aware of `.keys()` but as far as I can tell from searching the internet, there's no simple way of actually removing the keys from the dict.
Including handling non-existent keys, these appear to be the options:
```python
d = {k: True for k in d.keys() - set(to_remove)}  # slow, would make algorithm non-linear
```
```python
for k in to_remove:
    d.pop(k, None)
```
```python
for k in to_remove:
    if k in d:
        del d[k]
```
Now, none of these are as nice as a simple
```python
d -= set(to_remove)
```
so a dict still differs quite a lot from an ideal ordered set in that regard.

In practice, the current implementation
```python
self.twos -= set(P)
self.nodes_by_degree[2] -= set(P[1:])
```
would become something like
```python
self.twos.pop(v1, None)
for vi in P[1:]:
    self.twos.pop(vi, None)
    self.nodes_by_degree[2].pop(vi)
```
or even
```python
if v1 in self.twos:
    del self.twos[v1]
for vi in P[1:]:
    if vi in self.twos:
        del self.twos[vi]
    del self.nodes_by_degree[2][vi]
```
Now maybe I'm being nitpicky, but this feels needlessly tedious. Maybe #3912 will be given another thought one day.  
For now though, I can implement this if it's the recommended approach.