@maxtidev 2025-08-14
> `d = d.keys() - to_remove` which does create a new dict

Wouldn't this return a set though, losing the property of being ordered for subsequent operations?

> The best way I know to get an in-place ordered set subtraction using dicts for sets is your option 2

I implemented option 1 because of the slightly fewer lines and less indentation, but I agree that `d.pop(k, None)` can be unintuitive. Should I change it to option 2? 

> `d = {k: True for k in d.keys() - set(to_remove)}` is still linear, right?

Yes, I wasn't clear enough. The operation is linear in the sizes of `d` and `to_remove`  instead of linear in *just* the size of `to_remove`, making the "LinearTime" approximate maximum independent set algorithm in question not be linear in the graph size anymore, if we were to use it.