@rossbar 2025-08-14
Returning to the motivation for ordering - the output of the function is a set, so the output will inherently be unordered. Therefore I don't think there's any need to worry about ordering.