@maxtidev 2025-08-14
The benefit of having the sets be ordered (until the end of the algorithm) is that the algorithm is deterministic which wouldn't be the case for unordered sets.  
It's not as much about the order as it is about `twos.pop()` etc. always returning the same value for the same situation. Graphs that have strings as node identifiers could get different results on consecutive calls of `maximum_independent_set(G, method="lineartime")` if the sets remained unordered.  
The result gets turned into a set after extending the set to be maximal. Since no further set operations are applied after this, the order doesn't matter anymore. Before that though, it matters for predictable behavior / reproducible results.