@dschult 2025-08-15
> d = d.keys() - to_remove which does create a new dict
>
> Wouldn't this return a set though, losing the property of being ordered for subsequent operations?

Yes, you are right. This set subtraction idiom for dicts-as-sets provides a set, not a dict-as-set. Sorry for the confusion.

As for `pop` vs `del`, the `del` is faster (even with an `if`) but readability is more important except in an inner loop. And I don't think it will matter here whichever you choose.